% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.barplot.r
\name{Mixed.Barplot}
\alias{Mixed.Barplot}
\alias{mixed.barplot}
\title{Function to plot the mixed barplots of the frequencies of individuals with (1) and without (0) the targeted condition.}
\usage{
mixed.barplot(ref, test, position.legend = "topright")
}
\arguments{
\item{ref}{The reference standard. A column in a data frame or a vector indicating the classification by the reference test. The reference standard must be coded either as 0 (absence of the condition) or 1 (presence of the condition)}

\item{test}{The index test or test under evaluation. A column in a dataset or vector indicating the test results in a continuous scale.}

\item{position.legend}{the position of the legend. Default"'topright'. See also \code{\link{legend}}.}
}
\value{
No Value returned.
}
\description{
This plot function shows the frequencies of the two groups and their overlap in a single graph. This function is intended to be used for tests with 20 or less different test values
}
\details{
The graph shows the overlapping barplots of the frequencies of the observed test results of two groups and their overlap. Many tests of intermediate quality have a considerable overlap. This graph allows for the visual inspection of the frequencies and their overlap. The zero scores (0) represent the norm group, 1 represent the targeted group (abnorm).
For studying the overlap, the mixed density plot is recommended: \code{plotMD}.
}
\examples{
# A short test with 5 ordinal values
test0     = rep(1:5, times=c(33,6,6,11,2))
test1   = rep(1:5, times=c(3,2,2,11,33))
ref = c(rep(0, length(test0)), rep(1, length(test1)))
test = c(test0, test1)
mixed.barplot(ref, test, position.legend='top')
# please note that test score 4 has complete overlapping frequencies.
}
\seealso{
\code{\link{UI.ordinal}} for calculating possible Uncertain Intervals for short, ordened tests, \code{\link{plotMD}} for plotting mixed densities.
}

