% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.cutpoint.R
\name{quality.cutpoint}
\alias{quality.cutpoint}
\title{Function for the description of te qualities of one or two decision thresholds or cutpoint.}
\usage{
quality.cutpoint(ref, test, cutpoint, cutpoint.higher = NULL)
}
\arguments{
\item{ref}{The reference standard. A column in a data frame or a vector indicating the classification by the reference test. The reference standard must be coded either as 0 (absence of the condition) or 1 (presence of the condition)}

\item{test}{The index test or test under evaluation. A column in a dataset or vector indicating the test results in a continuous scale.}

\item{cutpoint}{The decision threshold of a dichotomization method, or the lower decision threshold of a trichotomization method.}

\item{cutpoint.higher}{(default = NULL). The upper decision threshold of a trichotomization method. When NULL, the test cores are dicotomized.}
}
\value{
{ A list of}
\describe{
  \item{table}{The confusion table of {diag x ref}, where diag is the diagnosis based on the test, when applying the cutpoint(s). The reference standard (ref) has categoires 0 and 1, while the diagnosis based on the test scores (diag) has categories 0 and 1 in the case of applying a single cut-point (dischotomization), and the categories 0, NA and 1 in the case of trichomization. In the case of the Uncertain Area trichomization method, the row NA shows the count of test scores within the Uncertain Area. When applying the trichimization method TG-ROC, the row NA shows the count of the test scores within the Intermediate Range. Table cell {0, 0} shows the True Negatives (TN), cell {0, 1} shows the False Negatives (FN), cell {1, 0} shows the False Positives (FP), and cell {1, 1} shows the True Positives (TP).}
  \item{cut}{The values of the cutpoint(s).}
  \item{indices}{A named vector, with the following statistics for the test-scores with diagnosis 0 or 1:
    \itemize{
      \item{prevalence: }{Diagnosable patients with the targeted condition / Total diagnosable sample = (TP+FN)/(TN+FP+FN+TP)}
      \item{correct.classification.rate (or Accuracy): }{(TP+TN)/(TN+FP+FN+TP)}
      \item{balance.correct.incorrect : }{(TP+TN)/(FP+FN)}
      \item{specificity: }{TN/(TN+FN)}
      \item{sensitivity: }{TP/(TP+FN)}
      \item{negative.predictive.value: }{TN/(TN+FN)}
      \item{positive.predictive.value: }{TP/(TN+FN)}
      \item{neg.likelihood.ratio: }{(1-sensitivity)/specificity}
      \item{pos.likelihood.ratio: }{sensitivity/(1-specificity)}
      
  }
}
}
}
\description{
This function can be used for both dichotomization (single cut-point) methods and for trichomization (two cut-points) methods. In the case of the Uncertain Area trichotomization method, it provides descriptive statitistics for the test scores outside the Uncertain Area. For the TG-ROC trichomization method it provides te descriptive statistics for TG-ROC's Valid Range.
}
\examples{
# A test with considerable overlap, resulting in a relatively large Uncertain Area
ref=c(rep(0,500), rep(1,500))
test=c(rnorm(500,0,1), rnorm(500,1,2))
ua = uncertain.area(ref, test) # with warning message!
quality.cutpoint(ref, test, ua[1], ua[2])
}

