% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UncertainInterval.R
\name{UncertainInterval}
\alias{UncertainInterval}
\title{Set of functions for the determination of an Uncertain Interval of test scores}
\description{
A collection of functions to determine a range of test scores that are inconclusive and do not allow a diagnosis (other than Uncertain) and to access its qualities.
}
\details{
See Index
}
\references{
Landsheer, J. A. (2016). Interval of Uncertainty: An Alternative Approach for the Determination of Decision Thresholds, with an Illustrative Application for the Prediction of Prostate Cancer. PloS One, 11(11), e0166007.
}
\seealso{
\code{\link{uncertain.interval}},  \code{\link{plotMD}},  \code{\link{get.intersection}},  \code{\link{quality.threshold}},  \code{\link{quality.threshold.uncertain}}
}

