% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMD.R
\name{plotMD}
\alias{plotMD}
\title{Function to plot the mixed densities of distributions of individuals with (1) and without (0) the targeted condition.}
\usage{
plotMD(ref, test, n.breaks = 20, subtitle = "")
}
\arguments{
\item{ref}{The reference standard. A column in a data frame or a vector indicating the classification by the reference test. The reference standard must be coded either as 0 (absence of the condition) or 1 (presence of the condition)}

\item{test}{The index test or test under evaluation. A column in a dataset or vector indicating the test results in a continuous scale.}

\item{n.breaks}{The number of breaks used to construct the histograms. This number has to be equal or lower than the discernable values in the test.}

\item{subtitle}{Optional subtitle}
}
\value{
No Value returned.
}
\description{
This plot function shows the two distributions and their overlap in a single graph.
}
\details{
The graph shows the two distributions and their overlap. Many tests of intermediate quality have a considerable overlap. Also, the distibutions as estimated by the \code{density} function, using the gaussian kernel is shown. The intersection is indicated by a vertical line. This graph allows the visual inspection of the two distributions, as well a visual inspection of the approximation of the \code{density}, based on the gaussian kernel. When the density estimation is way off, the standard estimation of the intersection will be incorrect, and another estimation has to be supplied.

The function \code{plotMD} can also be used for visual inspection of the Uncertain Area (see examples). Please note that the sensitivity and specificity values > .5 (including the default of .55) allows for some positive bias.
}
\examples{
# A test of intermediate quality
set.seed(1)
ref=c(rep(0,500), rep(1,500))
test=c(rnorm(500,0,1), rnorm(500,1,1.2))
plotMD(ref, test)
ua = uncertain.area(ref, test) # with warning message!
# Add lines to indicate Uncertain Area
abline(v=ua[1:2])
select=(test <= ua[2] & test >= ua[1])
# plot the mixed densities for the Uncertain Area
plotMD(ref[select], test[select])
}

