% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI.ordinal.R
\name{UI.ordinal}
\alias{UI.ordinal}
\title{Experimental! Function to calculate all possible uncertain intervals of ordinal test results of individuals with (1) and without (0) the targeted condition.}
\usage{
UI.ordinal(ref, test, adjust = 2, ...)
}
\arguments{
\item{ref}{The reference standard. A column in a data frame or a vector indicating the classification by the reference test. The reference standard must be coded either as 0 (absence of the condition) or 1 (presence of the condition)}

\item{test}{The index test or test under evaluation. A column in a dataset or vector indicating the test results in a continuous scale.}

\item{adjust}{The adjust factor that is transferred to the density function. Default value is 2.}

\item{...}{Further parameters that are transferred to the density function.}
}
\value{
List of values:
\describe{
  \item{$Youden}{A vector of statistics concerning the maximized Youden index:}
     \itemize{
      \item{max.Youden: }{The value of the Maximized Youden Index (= max(tpr - fpr)).}
      \item{threshold: }{The threshold associated with the Maximized Youden Index. Test values >= threshold indicate the targeted condition.}
      \item{Sp: }{The Specificity of the test when this threshold is applied}
      \item{Se: }{The Sensitivity of the test when this threshold is applied}
      \item{Acc: }{The Accuracy of the test when this threshold is applied}
      }
  \item{data.table}{A data.frame with the following columns:}
     \itemize{
      \item{test: }{The test scores.}
      \item{d0: }{The frequencies of the test scores of the norm group.}
      \item{d1: }{The frequencies of the test scores of the group with the targeted condition.}
      \item{tot: }{The total frequency of each test scores.}
     \item{TP: }{The number of True Positives when this test score is used as threshold.}
     \item{FP: }{The number of False Positives when this test score is used as threshold.}
     \item{fpr: }{The false positive rate when this test score is used as threshold.}
     \item{tpr: }{The true positive rate when this test score is used as threshold.}
     \item{Y: }{The Youden Index (= tpr - fpr) when this test score is used as threshold.}
     }
  \item{intersection}{The (rounded) intersection for the distributions of the two groups. Most often these distributions have no true point of intersection and the rounded intersection is an approximation. Often, this equals the Maximized Youden threshold (see Schisterman 2005). }
  \item{Uncertain.Interval}{Data frame with the statistics of all possible bounds of the uncertain interval. The columns are the following: }
    \itemize{
      \item{lowerbound: }{Lower bound of the possible uncertain interval.}
      \item{upperbound: }{Upper bound of the possible uncertain interval.}
      \item{UI.Sp: }{Specificity of the test scores between and including the lower and upper boundary. Closer to .5 is 'better', that is, more uncertain.}
      \item{UI.Se: }{Sensitivity of the test scores between and including the lower and upper boundary. Closer to .5 is 'better', that is, more uncertain.}
      \item{UI.ratio: }{The ratio between the number of patients in the uncertain area with and without the condition. Closer to one is better; (.45/.55) < UI.ratio < (.55/.45) or 0.82 < UI.ratio < 1.22 as a rule of fist.}
      \item{UI.n: }{Number of patients with test scores between and including the lower and upper boundary.}
      \item{MCI.Sp: }{Specificity of the more certain interval, i.e., the test scores lower than the lower boundary and higher than the upper boundary.}
      \item{MCI.Se: }{Sensitivity of the test scores lower than the lower boundary and higher than the upper boundary.}
      \item{MCI.Acc: }{Accuracy of the test scores lower than the lower boundary and higher than the upper boundary.}
      \item{MCI.n: }{Number of patients with test scores lower than the lower boundary and higher than the upper boundary.}
      \item{Loss: }{Loss of the test scores lower than the lower boundary and higher than the upper boundary. The total loss is the sum of the loss of the three areas:
      lower MCI: the summed number of false positives for each test score, divided by the number of persons that have received that test score.
      upper MCI: the summed number of false negatives, divided by the number of persons that have received that test score.
      uncertain interval: the sum of the absolute differences in the number of people out of the norm group d0 and the number of persons in the group with the targeted condition (d1) per test score, divided by the total number of persons.}
      }
  \item{candidates: }{Candidates with a loss lower than the Youden loss which might be considered for the Uncertain Interval }
  }
  @references{ Schisterman, E. F., Perkins, N. J., Liu, A., & Bondell, H. (2005). Optimal cut-point and its corresponding Youden Index to discriminate individuals using pooled blood samples. Epidemiology, 73-81.

  Landsheer, J. A. (2016). Interval of Uncertainty: An alternative approach for the determination of decision thresholds, with an illustrative application for the prediction of prostate cancer. PLOS One.
  }
}
\description{
This function calculates the uncertain interval of the test results of the two groups. This function is intended to be used for tests with 20 or less ordened test values
}
\details{
Warning: This function has not been sufficiently tested or studied.

Due to the limited possibilities of short scales, it is more difficult to determine a suitable uncertain interval. For any threshold determination, one needs a representative sample of sufficient size (200 or larger). If there are no test scores below the intersection in the candidate uncertain area, Sp of the Uncertain Interval (UI.Sp) is not available, while UI.Se equals 1.
The essential question is always whether the patients with the test scores inside the uncertain interval can be sufficiently distinguished.
Discussion of the first example (please run the code first): Visual inspection of the Mixed Barplot shows that distinguishing patients with and without the targeted condition is almost impossible for test scores 2, 3 and 4.
Sensitivity and Specificity of the uncertain interval should be not too far from .5. In the first example, the first interval (3:3) has no lower scores than the intersection (3), and therefore Ui.Sp is not available and UI.Se = 1. The UI.ratio indicates that the number of patients with and without the condition are equal in this interval. For these 110 patients, a diagnosis of uncertainty is probably the best choice.
The second interval (3:4) has an UI.Sp of .22, which is a large deviation from .5. In this slightly larger interval, the patients with a test score of 3 have a slightly larger probability to belong to the group without the condition. UI.Se is .8. UI.ratio is close to 1, which makes it a feasible candidate.
The third interval (2:4) has an UI.Sp of .35 and an UI.Se of .70 and an UI.ratio still close to one.
The other intervals show either Se or Sp that deviate strongly from .5, which makes them unsuitable choices.
Probably the easiest way to determine the uncertain interval is the interval with minimum loss. This is interval (2:4).
The Loss formula is (created by https://www.codecogs.com/latex/eqneditor.php):
\deqn{L =\frac{ \left (\sum_{i=l}^{u} \left |d0_{i}-d1_{i}  \right | + \sum_{i=u+1}^{h} d1_{i}+ \sum_{i=1}^{l-1}d0_{i}\right )}{N}}{ L = 1/N * (sum(abs(d0[u:l] - d1[u:l])) + sum(d1[1:(l-1)]) + sum(d0[(u+1):h]))}
where \emph{d0} represents the test scores of the norm group, \emph{d1} represents the test scores of the targeted patient group, \emph{l} is the lower limit of the uncertain interval, \emph{u} the upper limit, the first test score is enumerated 1 and the last test score is enumerated \emph{h}. \emph{N} is the total number of all persons with test scores.
 \itemize{
 \item{\eqn{\sum_{i=l}^{u} \left |d0_{i}-d1_{i}  \right |}{sum(abs(d0[u:l] - d1[u:l])}}{ is the loss in the uncertain interval, that is, the total deviation from equality.}
 \item{\eqn{\sum_{i=u+1}^{h} d1_{i}}{sum(d1[1:(l-1)])}}{ is the loss in the lower More Certain Interval, that is, the total of False Negatives, the number of patients with the targeted condition with a test score lower than \emph{l}, and}
 \item{\eqn{\sum_{i=u+1}^{h} d0_{i}}{sum(d0[(u+1):h])}}{ is the loss in the lower More Certain Interval, that is, the total of False Positives, the number of patients without the targeted condition with a test score higher than \emph{u}.}

The Loss is higher when the deviation from equality is higher in the uncertain area, higher when the number of False Negatives is higher, and higher when the number of False Positives is higher. The loss of a single threshold method equals 1 - its Accuracy.
In this example, the minimum Loss is found with interval (2:4). As this agrees with acceptable values for UI.Se, UI.Sp and UI.ratio, this seems the most suitable choice. In this case, the number of patients with test scores 2 to 4 are almost as likely to come from either population.
The remaining cases outside the uncertain interval (2:4) show high Accuracy, Specificity and Sensitivity.

}
}
\examples{
# A short test with 5 ordinal values
test0     = rep(1:5, times=c(165,14,16,55, 10)) # test results norm group
test1     = rep(1:5, times=c( 15,11,13,55,164)) # test results of patients
ref = c(rep(0, length(test0)), rep(1, length(test1)))
test = c(test0, test1)
table(ref, test)
plotMD(ref, test, breaks=c(min(test)-.5, seq(min(test), max(test), by=1)+.5),
  adjust=2) # visual inspection
UI.ordinal(ref, test) # determination of the uncertain interval
}
\seealso{
{  \code{\link{plotMD}} for plotting the mixed densities of the test values.
            \code{link{stats::density}} for the parameters of the density function. }
}

